First commit with the merge of the Vladimir+German+Official versions. Now, the TRF trace is deprecated, and all improvements should be written on the DIM trace. In further commits, all code that manages the SDDF trace will be erased progresively adopting the DIM (Paraver style trace) codification of equivalent semantics.
There have been modifications by Vladimir in the original DIM
trace, basically adding offsets for all threads of all tasks,
sorting each task per thread and adding the thread field on
the P2P communication primitives. This has to be documented.

The current version supports the codification of CPU ratios
using user events present on DIM trace. Paraver and Dimemas
will use same event encoding, and Dimemas blocks are discontinued.

Now it lacks support of generating PCF file, but it can be
copied using the '-pc' parameter.

Most part of the File IO simulation is not currently supported.
It should be analyzed and ported.

It has been added the skeleton to implement the new scalable
IO, as part of the HOPSA project. Every IO made by Dimemas
should use the new interface to be implemented.

+++ src/utils/mallocame.h
@@ -34,12 +34,14 @@
 
 #ifndef __mallocame_h
 #define __mallocame_h
+
 /**
  * External routines defined in file malloc.c
  **/
+
 extern void malloc_init(void);
-extern char *mallocame(int s);
-extern void freeame(char  *a, int s);
+extern char *mallocame(size_t s);
+extern void freeame(char  *a, size_t s);
 extern void malloc_end();
 
 #endif
